+++ app/component/front-page/NearbyTabLabelContainer.js
@@ -6,7 +6,7 @@ import some from 'lodash/some';
 import flatten from 'lodash/flatten';
 import config from '../../config';
 import NearbyTabLabel from './NearbyTabLabel';
-import StopListRoute from '../../relay-route/StopListRoute';
+import StopListRoute from '../../route/StopListRoute';
 
 const STOP_COUNT = 20;
